Add login/add whitelist logic
Logic added that adds user whitelist when phone first signs in (Note:
if app is uninstalled and reinstalled the device is recognized as a
new device and gets a corresponding new deviceID.)

DashboardFragment added which currently displays package names with
checkmarks that currently do nothing. Service must be added next and
local database must be created. Dashboard fragment will have the app
label rather than the package name, and will have a button to start
and stop the service.

Also when a user is already signed in and returns to the app, they will
be asked to enter their password before continuing.